+++ src/App.tsx
@@ -1,67 +1,26 @@
-import { FormEvent, useEffect, useState } from 'react'
+import { FormEvent, useState } from 'react'
 import './App.css'
 import Greeting from './components/Greeting'
 import Navbar from './components/Navbar'
 import Post from './components/Post'
-import { CreatePostDTO, PostDTO } from './types/dto'
-import axios from 'axios'
+import usePosts from './hooks/usePosts'
 
 function App() {
-  const [posts, setPosts] = useState<PostDTO[] | null>(null)
+  const { posts, isLoading, isPosting, createPost } = usePosts()
   const [newTitle, setNewTitle] = useState<string>('')
   const [newBody, setNewBody] = useState<string>('')
-  const [isLoading, setIsLoading] = useState<boolean>(false)
-  const [isPosting, setIsPosting] = useState<boolean>(false)
-
-  useEffect(() => {
-    const fetchData = async () => {
-      setIsLoading(true)
-      try {
-        // const res = await fetch('https://jsonplaceholder.typicode.com/posts')
-        // const data = await res.json()
-
-        // if (!res.ok) {
-        //   throw new Error('error')
-        // }
-        const res = await axios.get<PostDTO[]>('https://jsonplaceholder.typicode.com/posts')
-
-        setPosts(res.data)
-      } catch (err) {
-        console.error(err)
-      } finally {
-        setIsLoading(false)
-      }
-    }
-
-    fetchData()
-  }, [])
 
   const handleSubmit = async (e: FormEvent) => {
     e.preventDefault()
 
-    const newPostBody: CreatePostDTO = {
-      userId: Math.floor(Math.random() * 1000),
-      title: newTitle,
-      body: newBody,
-    }
-
-    setIsPosting(true)
     try {
-      // * axios post request
-      const res = await axios.post<PostDTO>('https://jsonplaceholder.typicode.com/posts', newPostBody, {
-        headers: { 'Content-Type': 'application/json' },
-      })
+      await createPost(newTitle, newBody)
 
-      console.log(res.data)
+      setNewTitle('')
+      setNewBody('')
     } catch (err) {
       console.error(err)
-    } finally {
-      setIsPosting(false)
     }
-
-    // * Clear form after set posts
-    setNewTitle('')
-    setNewBody('')
   }
 
   if (isLoading) return <h1>Loading...</h1>
